refactor: rename plugin transformer
