+++ packages/babel-plugin-transform-react/transformer.js
@@ -1,0 +1,1 @@
+module.exports = require("./src/transformer")
